+++ tests/extract/test_excel_custom_ui_extractor.py
@@ -26,7 +26,7 @@ class TestExcelCustomUiExtractor:
     def sut(self) -> Generator[ExcelCustomUiExtractor]:
         """Act first this tests."""
         common_folder = SettingsCommonFolder(
-            Path(Path.cwd(), "tests", "test.xlsm"), ".VBA"
+            Path(Path.cwd(), "tests", "test.xlsm"), ".VBA", include_extension=True
         )
         settings = SettingsFoldersHandleExcel(
             settings_common_folder=common_folder,
@@ -42,6 +42,25 @@ class TestExcelCustomUiExtractor:
     def test_exists_custom_ui_14_xml_file(self, sut: ExcelCustomUiExtractor) -> None:  # noqa: ARG002
         """Test that customUI14.xml file exists."""
         expected_file = Path(
-            Path.cwd(), "tests", "test.VBA", "customUI", "customUI14.xml"
+            Path.cwd(), "tests", "test.xlsm.VBA", "customUI", "customUI14.xml"
         )
         assert Path.is_file(expected_file)  # noqa: S101
+
+    def test_custom_ui_folder_not_created_when_no_custom_ui_files(self) -> None:
+        """Test that custom_ui_folder is not created
+        when no custom UI files are present.
+        """  # noqa: D205
+        workbook_path = Path("tests/extract/no_custom_ui.xlsm")
+        common_folder = SettingsCommonFolder(workbook_path, ".VBA")
+        settings = SettingsFoldersHandleExcel(
+            settings_common_folder=common_folder,
+            export_folder="",
+            custom_ui_folder="customUI",
+            code_folder="",
+        )
+        if settings.common_folder.exists():
+            shutil.rmtree(settings.common_folder)
+
+        ExcelCustomUiExtractor(settings)
+
+        assert not settings.custom_ui_folder.exists()  # noqa: S101